Added WriteHeader support for new nginx module.
Made webapp_str_t use uint32_len for lengths, added allocated int to prevent struct size issues.
Moved HTTP header calculation to nginx.

+++ plugins/time.lua
@@ -431,7 +431,7 @@ function M.toddate(x) T_int64_str(x)
 end
 
 ffi.cdef[[
-unsigned long long GetWebappTime();
+uint64_t GetWebappTime();
 ]]
 
 c = ffi.C
